Blocking
Add code to stop monsters either being placed in a wall, or all on the
same tile.

This is achieved by adding a 'blocks' property to the generic object
class. This allows game objects (monsters, items, etc) to have defined
whether or not they can be walked over or not.

E.g. You would want to be able to walk over a key or health potion, but
necessarily a monster.

A function was also added to check whether an item being placed in a
room, is going to be placed on tile that it can't occupy (e.g. a wall,
another monster, npc, the player, etc)

modified:   cccrawler.py