Turns
The player and monster take alternate turns of moving or attacking. To
do that we make use of the game state setup previously, to determine
when the player has moved so that the monsters/npcs can then take their
turn.

A debug message is displayed to show this happening.

A function was also added, to replace the movement on key press. This
determines whether or not the tile the player is moving to is
attackable. If it is attack it, otherwise just move to the tile
(assuming it's not a wall.....)

modified:   cccrawler.py